move tests to separate package, fix h3_ffi polygon methods, fix tests
