+++ h3_dart/test/h3_dart_vm_test.dart
@@ -1,3 +1,6 @@
+@TestOn('vm')
+library;
+
 import 'dart:io';
 
 import 'package:test/test.dart';
