+++ h3_ffi/pubspec.yaml
@@ -14,5 +14,5 @@ dependencies:
 
 dev_dependencies:
   ffigen: ^19.0.0
-  collection: # uses workspace's version
   test: # uses workspace's version
+  h3_test: # uses workspace's version
