+++ h3_test/lib/h3_test.dart
@@ -1,0 +1,2 @@
+export 'src/h3_tests.dart';
+export 'src/models.dart';
